# (PoC) Speedup freeciv-server cold start by 19% (branch S3_2)