@cazfi 2024-08-26
> freeciv-server already links to glib:

That's a derived dependency, not something we control ourselves, and server (and, e.g., sdl client) should be buildable without glib on systems where there is no such derived dependency.

But your work sounds interesting. And even if it turns out that we can't use the code, it's good to know that this is the place to optimize.