@michaelortmann 2024-08-26
instead of glib, `utility/genhash.c` could be used.

the optimization of the utf stuff we currently do would collide with current freeciv design. freeciv seems to abstract away / hide all utf16 iuc stff in `utility/support.c` so using any utf16  low level stuff in the nation ruleset loader didnt look like an option to me.

a solution could be adding special functions to `utility/support.c`, like, for example, one function to initialize one of the utf16 string compare values, and one function to compare the once initialized string to to a given one. but then i saw the thread mutex stuff. putting iuc mutex in ruleset.c? no. i scratched my head over adding a callback function, and the more solutions i came up the more it became clear, that there is no unintrusive and beautiful solution, as long as the code handling our data transform is separated between support/utility.c and the rest of freeciv, like server/ruleset.c because of our utf16 stuff.

thats when i got stuck and switched to the idea i showcased here, circumventing all the utf16 stuff, solving it entirely in ruleset.c instead of support.c.

looking from  anaother angle, freeciv stores utf8 strings everywhere, but then it has to convert it to utf16 just for string comparisons all the time. we cant directly compare such raw utf8 strings, because 2 different utf8 strings could be the same unicode points, but there is utf8 normalization and stuff. freeciv could store all utf8 strings in their normalized version.